Remove obex
Obex is supposed to run as user on the session bus. But what we have
now is a file server running as root when it should not need it.  It
is also patches that require maintenance.

Let's reintroduce this when running as user when user daemons are
enabled on snapd.