snap: add serial-port plug to hciattach and btattach
The hciattach and btattach tools need to access the serial port which the
bluetooth device is connected to. Both tools operate in similar way, they open
the serial port and issue a bunch of ioctls() that change the line discipline to
N_HCI, set the protocol and speed, which triggers the kernel to create a hciX
device for that port.

Fixes: #8

Signed-off-by: Maciek Borzecki <maciek.borzecki@gmail.com>